C program to check whether a number is armstrong or not
This is a c program to check whether a number entered by the user is armstrong or not.  I have included steps for finding the length of the number so that the user need'nt restrict the number of digits  to 3 . Proper commenting is provided for better understanding. 
Do share your valuable opinions and ideas regarding this program in the comment section.